Move commit passive unmount/mount to CommitWork
